refactor: 기본키 매핑 변경, 엔티티 PK 데이터 타입 변경 1. 기본키 매핑 IDENTITY로 변경(데이터베이스에 위임 - 기본 키를 가져오기 위해 insert를 수행한다) 2. H2, MySQL의 BIGINT형은 java.lang.Long과 매핑된다

+++ src/main/java/com/hello/kurly/users/domain/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.math.BigInteger;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, BigInteger> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
   Optional<User> findByNickname(String nickname);
